// frontend/src/types/generated/mobile/conectaPROMobileAPI.schemas.ts
export type DeviceTokenCreateDeviceId = string | null;

export type DeviceTokenCreateDeviceName = string | null;

export type DeviceTokenCreateDeviceModel = string | null;

export type DeviceTokenCreateOsVersion = string | null;

export type DeviceTokenCreateAppVersion = string | null;

export type DeviceTokenCreateAppBuild = string | null;

export type DeviceTokenCreateTimezone = string | null;

/**
 * Schema para criar/registrar token de dispositivo.
 */
export interface DeviceTokenCreate {
  /**
   * Token FCM/APNs
   * @minLength 10
   * @maxLength 500
   */
  token: string;
  /** android, ios, web */
  platform: string;
  device_id?: DeviceTokenCreateDeviceId;
  device_name?: DeviceTokenCreateDeviceName;
  device_model?: DeviceTokenCreateDeviceModel;
  os_version?: DeviceTokenCreateOsVersion;
  app_version?: DeviceTokenCreateAppVersion;
  app_build?: DeviceTokenCreateAppBuild;
  /** @maxLength 10 */
  locale?: string;
  timezone?: DeviceTokenCreateTimezone;
}

export type DeviceTokenResponseDeviceId = string | null;

export type DeviceTokenResponseDeviceName = string | null;

export type DeviceTokenResponseDeviceModel = string | null;

export type DeviceTokenResponseOsVersion = string | null;

export type DeviceTokenResponseAppVersion = string | null;

export type DeviceTokenResponseLocale = string | null;

export type DeviceTokenResponseTimezone = string | null;

export type DeviceTokenResponseLastUsed = string | null;

/**
 * Response de token de dispositivo.
 */
export interface DeviceTokenResponse {
  id: string;
  user_id: number;
  platform: string;
  device_id?: DeviceTokenResponseDeviceId;
  device_name?: DeviceTokenResponseDeviceName;
  device_model?: DeviceTokenResponseDeviceModel;
  os_version?: DeviceTokenResponseOsVersion;
  app_version?: DeviceTokenResponseAppVersion;
  push_enabled?: boolean;
  is_active?: boolean;
  locale?: DeviceTokenResponseLocale;
  timezone?: DeviceTokenResponseTimezone;
  last_used?: DeviceTokenResponseLastUsed;
  created_at: string;
  updated_at: string;
}